fix-1 enhancement_333_sidebar_menu
