+++ src/features/SideBar/ui/SideBar.module.scss
