--- src/widgets/SideBar/model/types/types.ts
@@ -1,8 +1,0 @@
-import { ReactElement } from 'react'
-
-export interface ISideBar {
-  title?: string
-  isVisible?: boolean
-  onClick?: () => void
-  children?: ReactElement | JSX.Element | JSX.Element[]
-}
