+++ src/widgets/SideBarMenu/ui/SideBarMenu.tsx
@@ -1,13 +1,16 @@
 import { FC } from 'react'
 
+import SideBar from '@/features/SideBar'
+import { userData, noUserData } from '@/mockData/sideBarProfileData'
 import Heading, { HeadingType } from '@/shared/ui/Heading/Heading'
 import Link from '@/shared/ui/Link/Link'
-import SideBar from '@/widgets/SideBar'
-
-import { userData, noUserData } from '../model/data/data'
-import { ISideBarMenu } from '../model/types/types'
 
 import styles from './SideBarMenu.module.scss'
+
+export interface ISideBarMenu {
+  user: string
+  handleLogOut: () => void
+}
 
 /**
  * Компонент SideBarMenu раскрывающийся в бургер меню
@@ -29,19 +32,17 @@ const SideBarMenu: FC<ISideBarMenu> = ({ user, handleLogOut }) => {
       <ul className={styles.sideBar__list}>
         {data &&
           data.map((item, index) => (
-            <li key={index}>
-              <SideBar isVisible={true} title={item.title}>
-                <ul className={styles.sideBar__sublinks}>
-                  {item.routes?.map((el, i) => (
-                    <li key={i}>
-                      <Link to={el.route || '#'} className={styles.sideBar__sublink}>
-                        {el.subtitle}
-                      </Link>
-                    </li>
-                  ))}
-                </ul>
-              </SideBar>
-            </li>
+            <SideBar key={index} isVisible={true} title={item.title}>
+              <ul className={styles.sideBar__sublinks}>
+                {item.routes?.map((el, i) => (
+                  <li key={i}>
+                    <Link to={el.route || '#'} className={styles.sideBar__sublink}>
+                      {el.subtitle}
+                    </Link>
+                  </li>
+                ))}
+              </ul>
+            </SideBar>
           ))}
       </ul>
       {user && <SideBar onClick={handleLogOut} isVisible={false} title="Выход" />}
